Move poof code
