+++ JustTags/AppVM.swift
@@ -292,14 +292,6 @@ internal final class AppVM: NSObject, ObservableObject {
             .map { $0.makeKeyAndOrderFront(self) }
     }
     
-    private func doPoof(window: NSWindow) {
-        let loc = CGPoint(x: window.frame.maxX - 15, y: window.frame.maxY - 45)
-        let poofSize = 20.0
-        
-        NSAnimationEffect.poof
-            .show(centeredAt: loc, size: .init(width: poofSize, height: poofSize))
-    }
-    
 }
 
 extension AppVM: NSWindowDelegate {
