--- JustTags/Globals.swift
@@ -1,10 +1,0 @@
-//
-//  Globals.swift
-//  JustTags
-//
-//  Created by Yurii Zadoianchuk on 30/04/2022.
-//
-
-import SwiftUI
-
-internal let diffBackground: Color = .blue.opacity(0.3)
